Delete+Add/Edit

+++ Project/PropertyEditor/PropertyEditor/Models/PropertyData.cs
@@ -14,25 +14,25 @@ namespace PropertyEditor.Models
                 new Property
                 {
                     Name = "width",
-                    Type = property_type.integer_type,
+                    Type = (int)property_type.integer_type,
                     Value = "800"
                 },
                 new Property
                 {
                     Name = "height",
-                    Type = property_type.integer_type,
+                    Type = (int)property_type.integer_type,
                     Value = "600"
                 },
                 new Property
                 {
                     Name = "title",
-                    Type = property_type.string_type,
+                    Type = (int)property_type.string_type,
                     Value = "GuruPoint project"
                 },
                 new Property
                 {
                     Name = "subtitle",
-                    Type = property_type.string_type,
+                    Type = (int)property_type.string_type,
                     Value = "Demo task"
                 }
             }.ForEach(a => context.Properties.Add(a));
